requirements.txt
cryptography==37.0.4
Flask==2.2.0
PyJWT==2.4.0
pyOpenSSL==22.0.0
PyYAML==6.0
Werkzeug==2.2.1
gunicorn==20.1.0
pytest==7.1.2
pytest-testdox==3.0.1
